Refactor DatabaseSystem into modular repository architecture
