+++ packages/server/src/db/repositories/index.ts
@@ -1,0 +1,20 @@
+/**
+ * Repository Module - Barrel Export
+ *
+ * Exports all database repositories for easy import.
+ * Each repository handles a specific domain of database operations.
+ *
+ * Usage:
+ * ```typescript
+ * import { PlayerRepository, InventoryRepository } from './database/repositories';
+ * ```
+ */
+
+export { BaseRepository } from "./BaseRepository";
+export { CharacterRepository } from "./CharacterRepository";
+export { PlayerRepository } from "./PlayerRepository";
+export { InventoryRepository } from "./InventoryRepository";
+export { EquipmentRepository } from "./EquipmentRepository";
+export { SessionRepository } from "./SessionRepository";
+export { WorldChunkRepository } from "./WorldChunkRepository";
+export { NPCKillRepository } from "./NPCKillRepository";
